Remove extra build step from dockerfile
When trying to deploy a new version of gotrue with Marcus, we noticed
the `/health` endpoint wasn't returning version data. After some digging
Marcus found that we're doing double builds in the Docker image, one via
`make build` and then a `go build` after that. The `go build` one
doesn't havee the version info and is redundant given the step before
it.

This change removes the `go build` step so we rely in the makefile one.