fix: convert `string` -> `*string` for AAL and AMR (#785)
## What kind of change does this PR introduce?

There was a brief incident last week where users with pre-MFA sessions
would run into errors when attempting to sign in. This was because past
sessions would have an empty `AAL` claims field(thus null) while the
Sessions model was expecting a string. This resulted in an error as a
null was being coerced into a string.

## What is the current behavior?
Both models use `string`

## What is the new behavior?

We modify both the AAL and AMR fields to accept `*string` in case there
are future instances which have null values for some reason

## Additional context


How this was tested:
1. Revert to `14068351d09a441edb738e62b591826406d589b6` (pre-MFA)
2. Create a few sessions by making a few signups, check DB to ensure
they have null field
3. Pull to come back to the current version and apply existing
migrations.
4. Attempt to sign up, sign in, get user, and then make a full
enrolment.

Co-authored-by: joel@joellee.org <joel@joellee.org>
Co-authored-by: Stojan Dimitrovski <sdimitrovski@gmail.com>