fix: create identity for invited user (#895)
## What kind of change does this PR introduce?
* Previously, inviting a user won't lead to an identity being created.
This creates a bug when the following steps are executed:
  1. Invite a new user
  2. Accept the invite via email
3. Try to sign-in with an oauth provider (i.e google) with the same
email used in (1)
This leads to an error saving the user in (3) because
`DetermineAccountLinking` will not be able to find any similar
identities as shown in:
https://github.com/supabase/gotrue/blob/73ce62873b5db50d442be91d240f3059e89b6f42/models/linking.go#L73-L79
Thus, this returns the decision to create a new account. However,
attempting to create a new user will result in a unique constraint
violation since a user with that email already exists.