feat: support for whatsapp as a channel for sending OTPs  (#981)
## What kind of change does this PR introduce?

We wish to support WhatsApp as a channel for sending OTPs due to the
added security provided and more. For this first pass, we will only
support Twilio.

Companion PR:
- Frontend:  https://github.com/supabase/gotrue-js/pull/616

For now, we only support Twilio API. We can slowly ungate more providers
as needed. There should be no change to the API guarantees otherwise -
if the user does not specify `whatsapp` the API will default to SMS.


I've excluded `reauthenticate` and `resend` methods from whatsapp
support as we don't currently use them in the client library. I'm not
sure whether we plan on supporting them long term so have left them with
SMS as default in order to reduce complexity. Let me know if anyone
feels otherwise and we can consider amending.

For now, we also expose the `updateUser` endpoint only on the API level
and not at the client library so as to minimize potential breaking
changes

**Uncertainties**
- For Supabase, I'll need to check if users can (and want to) use the
same Whatsapp and SMS number for sending via Twilio. Otherwise, we may
need to introduce additional configuration variables as a last resort.

Testing
- This was manually test and and an existing check for testing channel
parameters was added

---------

Co-authored-by: joel@joellee.org <joel@joellee.org>
Co-authored-by: Stojan Dimitrovski <sdimitrovski@gmail.com>