fix: log correct referer value (#1178)
## What kind of change does this PR introduce?
* Currently, we always log
[`r.Referer()`](https://cs.opensource.google/go/go/+/refs/tags/go1.20.5:src/net/http/request.go;l=454)
in the gotrue logs. However, this assumes that the referrer is always
sent in the header which is false.
* Move the functions to get the referrer and validate the referrer to
the `utilities` package so it can be used across the `api` and
`observability` packages
* Removed `getRedirectURLOrReferrer` because it's basically a repeat of
`getReferrer`. You would still need to parse the `redirect_to` before
calling `getRedirectURLOrReferrer` and `getReferrer` handles the parsing
+ validation for us already.