feat: retry concurrent refresh token attempts (#1202)
With #1190 each refresh token request was serialized with other such
requests. However, in cases where there's a not-insignificant number of
parallel refresh token requests for the same refresh token, a situation
arises where a significant number of connections are opened against the
database that perform little work. Furthermore, if GoTrue is configured
with a bounded connection pool, this could lead to significantly reduced
performance of the server.

With these changes, when `SELECT ... FOR UPDATE` is used on the refresh
token and session row, it is now done using the `SKIP LOCKED` clause. If
one or both of those rows are already locked, GoTrue will immediately
release the database connection, wait 10-30 ms and try to refresh the
token again. If after 5s the lock cannot be acquired, a HTTP 409
Conflict error will be returned. Clients should consider waiting at
least a few seconds before retrying on this error.