feat: add email rate limit breach metric (#1208)
## What kind of change does this PR introduce?

This PR aims to expose a set of email rate limit metrics as a Prometheus
metric that can then be consumed by an alerting system like Prometheus.
When a rate limit is triggered we increment a counter. This can then be
passed to a monitoring system such as Prometheus alert manager which can
fire off a notification (or similar) when a threshold (say 5 occurrences
in an hour) has been breached.

Extends: #1213 


The presence of the metric was tested via using the default
`prometheus.yml` file that comes on download. To test that the rate
limit is firing, we decreased `GOTRUE_RATE_LIMIT_EMAIL_SENT="5"` to 5
and ran `ab` against the endpoint like

`ab -p mass_signup.txt -T application/json -c 10 -n 50
http://localhost:9999/otp`

---------

Co-authored-by: joel@joellee.org <joel@joellee.org>