fix: allow transactions to be committed while returning a custom error (#1310)
## What kind of change does this PR introduce?
* This PR introduces a new error type `CommitWithError` that allows one
to commit a transaction but also return an error.
* This is useful in situations where
`GOTRUE_MAILER_ALLOW_UNVERIFIED_EMAIL_SIGN_INS="false"` since oauth
users with an unverified email will require email confirmation before
being allowed to sign-in. If the transaction doesn't get committed, the
new user doesn't get created and the email confirmation sent out will
not be mapped to a user in the database.