ci: add explicit release version in docker build (#1412)
Publishing builds the image, but the Makefile uses `git describe --tags`
for the version if the `RELEASE_VERSION` env var is not defined. And
this is an issue because the `.git` directory is not included in the
image (and it shouldn't be). Furthermore, the describe tags does not
reflect the proper version name.