fix: move all EmailActionTypes to mailer package (#1510)
## What kind of change does this PR introduce?

Moves the mail related refactors from #1496 into a new PR so the hook
related PR is easier to review. This PR moves all EmailActionTypes to
mailer package to establish an explicit link between mailer and the
packages it is used in.

The changes are cosmetic and should not affect underlying functionality.