feat: add authorized email address support (#1757)
Adds support for authorized email addresses, useful when needing to
restrict email sending to a handful of email addresses. In the Supabase
platform use case, this can be used to allow sending of emails on new
projects only to the project's owners or developers, reducing email
abuse.

To enable it, specify `GOTRUE_EXTERNAL_EMAIL_AUTHORIZED_ADDRESSES` as a
comma-delimited string of email addresses. The addresses should be
lowercased and without labels.

Labels are supported so emails will be sent to `someone+test1@gmail.com`
and `someone+test2@gmail.com` if and only if the `someone@gmail.com`
address is added in the authorized list.

Not a substitute for blocklists!