feat: add `x-sb-error-code` header, show error code in logs (#1765)
Adds the `x-sb-error-code` header to non-2XX responses if the error has
an error code determined. This is picked up by the request logger which
includes it in the request completed log.

Furthermore, the same header can be picked up by the API gateway
(Supabase world) without having to parse the response to include the
error code in request logs too.