fix: fallback on btree indexes when hash is unavailable (#1856)
## What kind of change does this PR introduce?
Edits an existing migration to fallback onto btree indexes when hash
indexes are unavailable.

This change allows auth to be compatible with the orioledb storage
engine. At time of writing orioledb only support btree indexes. That
issue is expected to be addressed by the end of Q1 - 2025. Once that
support is added, both heap and orioledb projects will both use the code
path for hash indexes. hash indexes are preferred because they are
[safer from timing
attacks](https://dba.stackexchange.com/questions/285739/prevent-timing-attacks-in-postgres)

A reminder has been set for May 1st for us to
- remove the fork in this migration
- identify and update the index type (from btree to hash) on any
orioledb-alpha projects that are still in existence