ci: introduce dogfooding check on release (#1909)
Adds a dogfooding check on release, i.e. making sure that the RC version
being released has already been out on Supabase production & staging.

It works by adding a mandatory dogfooding check on every PR. If the PR
is one for Release Please, upon giving approval to merge, the check will
re-run ensuring that the latest RC is in fact deployed. If it is not,
the check will fail, blocking the merge.