feat: experimental own linking domains per provider (#2119)
Experimental feature internal to Supabase. Don't use outside of
Supabase.

Adds the ability to specify providers e.g. `google`, `apple`, whose
identities will not participate in the `default` email-similarity
identity linking algorithm but be on its own. You can do this by setting

```
GOTRUE_EXPERIMENTAL_PROVIDERS_WITH_OWN_LINKING_DOMAIN="google,apple,..."
```

This has the effect of setting the `is_sso_user` column on `auth.users`
to true, which is the only mechanism that allows the same email address
to appear in different linking domains. The column cannot be renamed as
that requires re-creating a partial index, which is not on the table for
this change. It will be revised in the future.