feat: add phone to sms webhook payload (#2160)
## What kind of change does this PR introduce?

This PR adds the phone number to the SendSMS webhook, when a phone
factor is challenged or a phone confirmation is required.

## What is the current behavior?

Currently, the SMS webhook only sends the `user_id` and the `otp_code`.
However, we have no way of knowing which factor (device) was used in
case the send SMS webhook is used for MFA (as opposed to phone
confirmation, which uses the user's phone number), and there is more
than one device enrolled for a single user.

## What is the new behavior?

The webhook payload contains `phone`, which is either the user's phone
number for phone confirmation or the factor phone number for MFA with a
phone factor.

## Additional context

Add any other context or screenshots.